[fs] ext4: cleanup bh release code in ext4_ind_remove_space()
Message-id: <20190517091449.6738-76-lczerner@redhat.com>
Patchwork-id: 256607
O-Subject: [RHEL8.1 fs PATCH 75/89] ext4: cleanup bh release code in ext4_ind_remove_space()
Bugzilla: 1698815
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>
RH-Acked-by: Bill O'Donnell <billodo@redhat.com>

From: "zhangyi (F)" <yi.zhang@huawei.com>

BZ 1698815
https://bugzilla.redhat.com/show_bug.cgi?id=1698815

BREW 21544584
https://brewweb.devel.redhat.com/taskinfo?taskID=21544584

commit 5e86bdda41534e17621d5a071b294943cae4376e

Currently, we are releasing the indirect buffer where we are done with
it in ext4_ind_remove_space(), so we can see the brelse() and
BUFFER_TRACE() everywhere.  It seems fragile and hard to read, and we
may probably forget to release the buffer some day.  This patch cleans
up the code by putting of the code which releases the buffers to the
end of the function.

Signed-off-by: zhangyi (F) <yi.zhang@huawei.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Reviewed-by: Jan Kara <jack@suse.cz>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>