[fs] Revert "ext4: use ext4_write_inode() when fsyncing w/o a journal"
Message-id: <20190517091449.6738-80-lczerner@redhat.com>
Patchwork-id: 256594
O-Subject: [RHEL8.1 fs PATCH 79/89] Revert "ext4: use ext4_write_inode() when fsyncing w/o a journal"
Bugzilla: 1698815
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>
RH-Acked-by: Bill O'Donnell <billodo@redhat.com>

From: Theodore Ts'o <tytso@mit.edu>

BZ 1698815
https://bugzilla.redhat.com/show_bug.cgi?id=1698815

BREW 21698455
https://brewweb.devel.redhat.com/taskinfo?taskID=21698455

commit 8fdd60f2ae3682caf2a7258626abc21eb4711892

This reverts commit ad211f3e94b314a910d4af03178a0b52a7d1ee0a.

As Jan Kara pointed out, this change was unsafe since it means we lose
the call to sync_mapping_buffers() in the nojournal case.  The
original point of the commit was avoid taking the inode mutex (since
it causes a lockdep warning in generic/113); but we need the mutex in
order to call sync_mapping_buffers().

The real fix to this problem was discussed here:

https://lore.kernel.org/lkml/20181025150540.259281-4-bvanassche@acm.org

The proposed patch was to fix a syzbot complaint, but the problem can
also demonstrated via "kvm-xfstests -c nojournal generic/113".
Multiple solutions were discused in the e-mail thread, but none have
landed in the kernel as of this writing.  Anyway, commit
ad211f3e94b314 is absolutely the wrong way to suppress the lockdep, so
revert it.

Fixes: ad211f3e94b314a910d4af03178a0b52a7d1ee0a ("ext4: use ext4_write_inode() when fsyncing w/o a journal")
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Reported: Jan Kara <jack@suse.cz>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>