[sound] ALSA: ASoC: Intel: bytcr_rt5651: Change default input map from in2 to in1
Message-id: <15580872947136052-111-patchset-jkysela@redhat.com>
Patchwork-id: 256747
O-Subject: [RHEL 8.1 PATCH 0110/2119] ALSA: ASoC: Intel: bytcr_rt5651: Change default input map from in2 to in1
Bugzilla: 1683051
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Further testing on all 6 model x86 tablets with a rt5651 which I have
access to for testing has shown that their single (mono) microphone is
connected to both IN1 *and* IN2.

The previous default mapping of IN2 was based on testing on the same 6
tablets, where the internal mic works fine with a mapping of IN2. But it
works fine too with a mapping of IN1.

This commit changes the default input mapping to to use IN1 instead of
IN2, to match the mapping used for the other mono devices in the DMI quirk
table. So that we need less different mappings.

The same change is made to the Chuwi Vi8 Plus quirks, which is one of the
6 models tested.

This is a preparation patch for simplifying the maps in a follow-up commit.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 10876d24eb40c6bfaa0aabd97e3e143258176c53)
Bugzilla: 1683051
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>