[sound] ALSA: seq: virmidi: Offload the output event processing
Message-id: <15580872947136052-375-patchset-jkysela@redhat.com>
Patchwork-id: 256939
O-Subject: [RHEL 8.1 PATCH 0374/2119] ALSA: seq: virmidi: Offload the output event processing
Bugzilla: 1683051
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The virmidi sequencer stuff tries to translate the rawmidi bytes to
sequencer events and deliver the packets at trigger callback.  The
amount of the whole process of these translations and deliveries
depends on the incoming rawmidi bytes, and we have no limit for that;
this was the cause of a CPU soft lockup that had been reported and
fixed recently.

Although we've fixed the soft lockup by putting the temporary unlock
and cond_resched(), it's rather a quick band aid.  In this patch,
meanwhile, the event parsing and delivery process is offloaded to a
dedicated work, and the trigger callback just kicks it off.  It has
three merits, at least:

- The processing is always done in a sleepable context, which can
  assure the event delivery with non-atomic flag without hackish
  is_atomic() usage.

- Other relevant codes can be simplified, reducing the lines

- It makes me happier

Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit f7debfe54090d1a1c38e1f070be20d83bb70a8e0)
Bugzilla: 1683051
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>