[sound] ALSA: HD-Audio: SKL+: force HDaudio legacy or SKL+ driver selection
Message-id: <15580872947136052-1119-patchset-jkysela@redhat.com>
Patchwork-id: 257744
O-Subject: [RHEL 8.1 PATCH 1118/2119] ALSA: HD-Audio: SKL+: force HDaudio legacy or SKL+ driver selection
Bugzilla: 1683051
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

For HDaudio and Skylake drivers, add module parameter "pci_binding"

When pci_binding == 0 (AUTO), the PCI class/subclass info is used to
select drivers based on the presence of the DSP.

pci_binding == 1 (LEGACY) forces the use of the HDAudio legacy driver,
even if the DSP is present.

pci_binding == 2 (ASOC) forces the use of the ASOC driver. The
information on the DSP presence is bypassed.

The value for the module parameter needs to be identical for both
drivers. This parameter is intended as a back-up solution if the
automatic detection fails or when the DSP usage fails. Such cases
should be reported on the alsa-devel mailing list for analysis.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit d82b51c855a20eb456ac09f2f40ea98312373263)
Bugzilla: 1683051
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>