[sound] ALSA: ASoC: tlv320aic32x4: Dynamically Determine Clocking
Message-id: <15580872947136052-1767-patchset-jkysela@redhat.com>
Patchwork-id: 258626
O-Subject: [RHEL 8.1 PATCH 1766/2119] ALSA: ASoC: tlv320aic32x4: Dynamically Determine Clocking
Bugzilla: 1683051
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The existing code uses a static lookup table to determine the
settings of the various clock devices on board the chip.  This is
limiting in a couple of ways.  First, this doesn't allow for any
master clock rates other than the three that have been
precalculated.  Additionally, new sample rates are difficult to
add to the table.  Witness that the chip is capable of 192000 Hz
sampling, but it is not provided by this driver.  Last, if the
driver is clocked by something that isn't a crystal, the
upstream clock may not be able to achieve exactly the rate
requested in the driver.  This will mean that clocking will be
slightly off for the sampling clock or that it won't work at all.

This patch determines the settings for all of the clocks at
runtime considering the real conditions of the clocks in the
system.  The rules for the clocks are in TI's SLAA557 application
guide on pages 37, 51 and 77.

Signed-off-by: Annaliese McDermond <nh6z@nh6z.net>
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 96c3bb00239de4fb5f4ddca42c1f90d6d9b3c697)
Bugzilla: 1683051
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>