[sound] ALSA: core: Don't refer to snd_cards array directly
Message-id: <15580872947136052-1913-patchset-jkysela@redhat.com>
Patchwork-id: 258475
O-Subject: [RHEL 8.1 PATCH 1912/2119] ALSA: core: Don't refer to snd_cards array directly
Bugzilla: 1683051
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The snd_cards[] array holds the card pointers that have been currently
registered, and it's exported for the external modules that may need
to refer a card object.  But accessing to this array can be racy
against the driver probe or removal, as the card registration or free
may happen concurrently.

This patch gets rid of the direct access to snd_cards[] array and
provides a helper function to give the card object from the index
number with a refcount management.  Then the caller can access to the
given card object safely, and releases it via snd_card_unref().

While we're at it, add a proper comment to snd_card_unref() and make
it an inlined function for type-safety, too.

Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit f4fa968950aef7efb98a0899bb0132405e2edaf1)
Bugzilla: 1683051
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>