[kernel] perf/core: Add sanity check to deal with pinned event failure
Message-id: <8a5fa8bb4ee2a0b06c9b91fc4e729d9b6ac9d397.1558604935.git.mpetlan@redhat.com>
Patchwork-id: 259752
O-Subject: [RHEL8.1 BZ1709242 025/190] perf/core: Add sanity check to deal with pinned event failure
Bugzilla: 1709242
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1709242
https://bugzilla.redhat.com/show_bug.cgi?id=1709242

upstream
========
commit befb1b3c2703897c5b8ffb0044dc5d0e5f27c5d7
Author: Reinette Chatre <reinette.chatre@intel.com>
Date: Wed Sep 19 10:29:06 2018 -0700

description
===========
It is possible that a failure can occur during the scheduling of a
pinned event. The initial portion of perf_event_read_local() contains
the various error checks an event should pass before it can be
considered valid. Ensure that the potential scheduling failure
of a pinned event is checked for and have a credible error.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>