[tools] perf hists: Clarify callchain disabling when available
Message-id: <7722474575671d6df0dde69b42829713110e238d.1558604935.git.mpetlan@redhat.com>
Patchwork-id: 259761
O-Subject: [RHEL8.1 BZ1709242 033/190] perf hists: Clarify callchain disabling when available
Bugzilla: 1709242
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1709242
https://bugzilla.redhat.com/show_bug.cgi?id=1709242

upstream
========
commit e9de7e2f7e22989fefc760cf0920062c58b2b2b1
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Jun 20 15:58:20 2018 -0300

description
===========
We want to allow having mixed events with/without callchains, not
using a global flag to show callchains, but allowing supressing
callchains when they are present.

So invert the logic of the last parameter to hists__fprint() to
that effect.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>