[tools] perf parser: Improve error message for PMU address filters
Message-id: <d1ded790ce1e77ec86d3fb077e3240cf2e093ac7.1558604935.git.mpetlan@redhat.com>
Patchwork-id: 259856
O-Subject: [RHEL8.1 BZ1709242 154/190] perf parser: Improve error message for PMU address filters
Bugzilla: 1709242
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1709242
https://bugzilla.redhat.com/show_bug.cgi?id=1709242

upstream
========
commit 49836f7811f383d8613661fff110ce74f4710d52
Author: Jack Henschel <jackdev@mailbox.org>
Date: Wed Jul 4 14:13:45 2018 +0200

description
===========
This is the second version of a patch that improves the error message of
the perf events parser when the PMU hardware does not support address
filters.

Previously, the perf returned the following error:

  $ perf record -e intel_pt// --filter 'filter sys_write'
  --filter option should follow a -e tracepoint or HW tracer option

This implies there is some syntax error present in the command line,
which is not true. Rather, notify the user that the CPU does not have
support for this feature.

For example, Intel chips based on the Broadwell micro-archticture have
the Intel PT PMU, but do not support address filtering.

Now, perf prints the following error message:

  $ perf record -e intel_pt// --filter 'filter sys_write'
  This CPU does not support address filtering

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>