[tools] perf util: Fix bad memory access in trace info.
Message-id: <a8462fccf3f71d4a6586adb8d223337e88342584.1558604935.git.mpetlan@redhat.com>
Patchwork-id: 259885
O-Subject: [RHEL8.1 BZ1709242 164/190] perf util: Fix bad memory access in trace info.
Bugzilla: 1709242
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1709242
https://bugzilla.redhat.com/show_bug.cgi?id=1709242

upstream
========
commit a72f64261359b7451f8478f2a2bf357b4e6c757f
Author: Chris Phlipot <cphlipot0@gmail.com>
Date: Tue Aug 28 23:19:54 2018 -0700

description
===========
In the write to the output_fd in the error condition of
record_saved_cmdline(), we are writing 8 bytes from a memory location on
the stack that contains a primitive that is only 4 bytes in size.
Change the primitive to 8 bytes in size to match the size of the write
in order to avoid reading unknown memory from the stack.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>