[tools] tools headers uapi: Sync kvm.h copy
Message-id: <bbcc385e7e3bda78a4f2c4439c12902484f9ad11.1558604935.git.mpetlan@redhat.com>
Patchwork-id: 259883
O-Subject: [RHEL8.1 BZ1709242 181/190] tools headers uapi: Sync kvm.h copy
Bugzilla: 1709242
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1709242
https://bugzilla.redhat.com/show_bug.cgi?id=1709242

upstream
========
commit 25fe15e54fe5e15b4963fe101f7cd8bad4f11393
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Oct 8 12:09:14 2018 -0300

description
===========
To pick up the changes introduced in:

  6fbbde9a1969 ("KVM: x86: Control guest reads of MSR_PLATFORM_INFO")

That is not yet used in tools such as 'perf trace'.

The type of the change in this file, a simple integer parameter to the
KVM_CHECK_EXTENSION ioctl should be easier to implement tho, adding to
the libbeauty TODO list.

This silences this perf build warning:

  Warning: Kernel ABI header at 'tools/include/uapi/linux/kvm.h' differs from latest version at 'include/uapi/linux/kvm.h'
  diff -u tools/include/uapi/linux/kvm.h include/uapi/linux/kvm.h

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>