[drm] drm/i915: Fix ONLY icl+ max plane width
Message-id: <20190530013147.11438-90-lyude@redhat.com>
Patchwork-id: 261274
O-Subject: [RHEL8 drm 089/117] drm/i915: Fix ONLY icl+ max plane width
Bugzilla: 1707312 1688524 1685552 1685427 1665631 1665088 1664615 1664609 1664189 1664177 1664175 1664104 1662017 1660302 1643842 1640840 1639990 1639436 1637606 1633845 1633843 1633842 1504286
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Ben Skeggs <bskeggs@redhat.com>
RH-Acked-by: Jerome Glisse <jglisse@redhat.com>
RH-Acked-by: Dave Airlie <airlied@redhat.com>

Bugzilla: 1665631,1639436,1633843,1665088,1660302,1504286,1664609,1633842,1707312,1664189,1639990,1633845,1664175,1664615,1685552,1662017,1640840,1688524,1664104,1685427,1637606,1643842,1664177
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=21914589
Upstream Status: RHEL only.

There was a rather important change upstream regarding watermarks and
plane widths for skl+, glk and icl:

372b9ffb5799 ("drm/i915: Fix skl+ max plane width")

Originally I was going to try to avoid pulling this in because of the
risky non-icl watermark changes it made, but since it looks like we're
getting the icl watermark fixes we asked Intel for when I first started
the DRM backport it's probably a bad idea to leave this out for icelake.

So, give ourselves the best of both worlds: intead of backporting
all of 372b9ffb5799, simply backport the changes it makes to icelake and
keep everything else as-is for Skylake.

This change will eventually be overwritten, either in zstream or by the
next DRM backport. Because this is based off an upstream patch, no
upstreaming is needed.

Signed-off-by: Lyude Paul <lyude@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>